+++ earl/model.py
